handle catch

+++ FileHaven_2.4/src/servlets/RecoverFileServlet.java
@@ -134,8 +134,8 @@ public class RecoverFileServlet extends HttpServlet {
 			}
 		}
 		catch(Exception ex){
-			fdb.remove();
-			frdb.remove();
+			if(fdb!=null)fdb.remove();
+			if(frdb!=null)frdb.remove();
 			session.setAttribute("info_line1", "Recover File Failed.");
 			session.setAttribute("info_line2", ex.getMessage());
 			getServletContext().getRequestDispatcher("/Information.jsp").forward(request,response);
