+++ FileHaven_2.4/src/servlets/StartDeleteFileServlet.java
@@ -50,7 +50,7 @@ public class StartDeleteFileServlet extends HttpServlet {
 		getServletContext().getRequestDispatcher("/DeleteFile.jsp").forward(request,response);
 		}
 		catch(Exception ex){
-			fdb.remove();
+			if(fdb!=null)fdb.remove();
 			session.setAttribute("info_line1", "File access denied.");
 			session.setAttribute("info_line2", ex.getMessage());
 			getServletContext().getRequestDispatcher("/Information.jsp").forward(request,response);
